fix: change deep4net data dtype to float32

+++ src/models/eeginception.py
@@ -23,7 +23,7 @@ class EEGInception(BaseModel):
             percentile_95 = np.percentile(np.abs(item["data"][i]), 95, axis=0, keepdims=True)
             item["data"][i] = item["data"][i] / percentile_95
 
-        data = np.zeros((self.hparams.n_chans, item["data"].shape[1]))
+        data = np.zeros((self.hparams.n_chans, item["data"].shape[1]), dtype=np.float32)
         data[:item["data"].shape[0]] = item["data"]
 
         return data, item["labels"].max()
